fix: renames directory switch to switchInput
To fix storybook build error due to typescript module name case sensitivity bug on mac
