+++ core/components/atoms/switchInput/__stories__/variants/Appearance.story.tsx
