+++ core/components/atoms/switchInput/index.tsx
